refactor: 로딩 스피너 배경 설정
