+++ app/(anon)/myPage/profile/edit/page.tsx
@@ -11,7 +11,7 @@ import { getToken } from "@/utils/getToken";
 import React, { useEffect, useState } from "react";
 import styled from "styled-components";
 import { ErrorMessage } from "../page.styles";
-import LoadingSpinner from "@/components/loadingSpinner/loadingSpineer";
+import LoadingSpinner from "@/components/loadingSpinner/loadingSpinner";
 
 const CustomProfileImageUploader = styled(ProfileImageUploader)`
     margin-top: 10px;
@@ -121,12 +121,9 @@ export default function ProfileEdit() {
         }
     };
 
-    if (isLoading) {
-        return <LoadingSpinner />;
-    }
-
     return (
         <>
+            {isLoading && <LoadingSpinner />}
             <Form onSubmit={handleEditProfile}>
                 <CustomProfileImageUploader
                     profileImage={imgSrc}
